+++ templates/mesh-columns-1.php
@@ -16,9 +16,13 @@
  * Add the ability to add markup before Mesh section
  */
 do_action( 'mesh_section_before' );
+
+$mesh_element_attributes = get_mesh_element_attributes();
+$mesh_element_attributes = apply_filters( 'mesh_row_attributes', $mesh_element_attributes );
+$mesh_element_attributes = implode( ' ', $mesh_element_attributes );
 ?>
 
-<section <?php post_class(); ?> <?php mesh_section_background(); ?>>
+<section <?php post_class(); ?> <?php mesh_section_background(); ?> <?php echo $mesh_element_attributes; ?>>
 	<?php
 	/**
 	 * Add the ability to add markup before Mesh row
@@ -31,14 +35,19 @@ do_action( 'mesh_section_before' );
 		$collapse_column_spacing = get_post_meta( get_the_ID(), '_mesh_collapse', true );
 
 		$row_class = ( ! empty( $collapse_column_spacing ) ) ? 'row collapse' : 'row';
+		$title_class = array( 'small-12', 'columns', 'title-row' );
+	    $title_class = apply_filters( 'mesh_section_title_css_classes', $title_class );
+	    $title_class = implode( ' ', $title_class );
 	?>
 
 	<div class="<?php echo esc_attr( $row_class ); ?>">
 		<?php if ( ! empty( $title_display ) && 'no block title' !== strtolower( get_the_title() ) ) : ?>
-			<div class="small-12 columns title-row">
+			<div class="<?php echo esc_attr( $title_class ); ?>">
 				<h2 class="entry-title"><?php the_title(); ?></h2>
 			</div>
 		<?php endif; ?>
+
+		<?php do_action( 'mesh_columns_before' ); ?>
 
 		<?php
 
@@ -46,8 +55,12 @@ do_action( 'mesh_section_before' );
 
 		if ( ! empty( $blocks ) ) :
 			foreach ( $blocks as $block ) :
+
+            $mesh_element_attributes = get_mesh_element_attributes( $block->ID );
+            $mesh_element_attributes = apply_filters( 'mesh_column_attributes', $mesh_element_attributes );
+            $mesh_element_attributes = implode( ' ', $mesh_element_attributes );
 			?>
-				<div <?php mesh_block_class( $block->ID ); ?> <?php mesh_section_background( $block->ID ); ?>>
+				<div <?php mesh_block_class( $block->ID ); ?> <?php mesh_section_background( $block->ID ); ?> <?php echo $mesh_element_attributes; ?>>
 					<?php if ( ! empty( $block->post_title ) && 'no column title' !== strtolower( $block->post_title ) ) : ?>
 						<h3 class="entry-subtitle"><?php echo esc_html( apply_filters( 'the_title', $block->post_title ) ); ?></h3>
 					<?php endif; ?>
@@ -59,6 +72,8 @@ do_action( 'mesh_section_before' );
 			endforeach;
 		endif;
 		?>
+
+		<?php do_action( 'mesh_columns_after' ); ?>
 	</div>
 
 	<?php